+++ api.js
@@ -4,7 +4,7 @@ const _ = require("lodash");
 import InfoExtract from "./server/infoextractor";
 import { RoomRequestType } from "./common/models/messages";
 const { getLogger } = require('./logger.js');
-const permissions = require("./server/permissions.js");
+const permissions = require("./common/permissions");
 import roommanager from "./server/roommanager";
 const { rateLimiter, handleRateLimit, setRateLimitHeaders } = require("./server/rate-limit");
 import { Role } from "./common/models/types";
